refactor: loading, header 컴포넌트 분리
