+++ src/components/loadingContainer/index.tsx
@@ -1,0 +1,7 @@
+import { Container } from "./index.style";
+
+const LoadingContainer = ({ children }: { children: JSX.Element }) => {
+  return <Container>{children}</Container>;
+};
+
+export default LoadingContainer;
