+++ src/components/mainHeader/index.style.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const Container = styled.section`
+  height: inherit;
+
+  display: flex;
+  justify-content: center;
+
+  font-size: 3rem;
+  color: ${({ theme }) => theme.color.grey_100};
+
+  .logo {
+    margin-top: 1.6rem;
+  }
+`;
