+++ src/pages/main/index.tsx
@@ -1,7 +1,7 @@
 import LoginModal from "../../components/modal/LoginModal";
 import { useState } from "react";
 import ModalContainer from "../../components/modal/ModalContainer";
-import MainHeader from "../../components/header/MainHeader";
+import MainHeader from "../../components/mainHeader";
 import BottomButtons from "./bottomButtons";
 import { Container, GuideText, Title, MainImage } from "./index.style";
