+++ src/pages/symptomType/index.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import ContentHeader from "../../components/header/ContentHeader";
+import ContentHeader from "../../components/contentHeader";
 import SymptomTypeComponent from "./symptomTypeComponent";
 import { symptom_type } from "../../store/symptom_type";
 import SymptomModal from "../../components/modal/SymptomModal";
